2ºAno/2ºSemestre/BD/BD_projeto_final_nuno_miguel/requirements.txt
flask
flask_jwt_extended
psycopg2-binary
